requirements.txt
Flask==2.2.2
gunicorn==20.1.0
tflite-runtime==2.14.0
Pillow==9.5.0
scipy==1.10.1
numpy==1.24.3